Minor Display bug fixed, adjusted tests to work with newest changes (CalculationOverview)

+++ FuzzyFightSimulator/Assets/Scripts/Characters/Character.cs
@@ -8,7 +8,7 @@ public class Character : MonoBehaviour
 
     [Header("General Information")]
     [SerializeField] public string CharacterName;
-    [SerializeField] public double TrainingHours;
+    [SerializeField] public float TrainingHours;
     private DefenseState _defenseState;
 
     [Header("Life Points")]
